requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
audioop-lts==0.2.2
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
fastapi==0.119.0
ffmpy==0.6.3
filelock==3.20.0
fsspec==2025.9.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.4
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
ruff==0.14.1
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
pytest==7.4.0
pytest-django==4.11.0
pytest-cov==7.0.0
factory-boy==3.3.0
faker==19.0.0
djangorestframework-simplejwt==5.5.1
